pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-options-analysis-tools"
version = "0.2.0"
description = "Graphical and statistical analysis for stocks, options, derivatives and finance."
readme = "README.md"
license = { text = "AGPL-3.0-or-later" }
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.0",
    "pandas>=2.0",
    "plotly>=5.18",
    "matplotlib>=3.8",
    "requests>=2.28",
    "beautifulsoup4>=4.12",
    "psutil>=5.9",
    "python-dateutil>=2.8",
    "sec-edgar-downloader>=5.0",
    "xlsxwriter>=3.1",
    "yfinance>=0.2.36",
    "openpyxl>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.4",
    "mypy>=1.8",
    "types-python-dateutil>=2.8",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests", "scripts"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"src/options_analysis/core.py" = ["N802"]
"src/sec_analysis/core.py" = ["N801"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
